.1、添加模型验证JSR303标准

+++ webapp-themlef/src/main/java/com/lufy/webappthemlef/controller/UserController.java
@@ -21,7 +21,7 @@ public class UserController {
     @RequestMapping("/{id}")
     public String getUser(@PathVariable Integer id, Model model) {
 
-        model.addAttribute("user", new User(id, "张三", 20, "中国广州"));
+        model.addAttribute("user", new User((long)id, "张三", 20, "中国广州"));
         return "/user/detail";
     }
 
@@ -36,7 +36,7 @@ public class UserController {
     public String listUser(ModelMap model) {
         List<User> userList = new ArrayList<User>();
         for (int i = 0; i < 10; i++) {
-            userList.add(new User(i, "张三" + i, 20 + i, "中国广州"));
+            userList.add(new User((long) i, "张三" + i, 20 + i, "中国广州"));
         }
 
         model.put("users", userList);
